rm: resources from BaseEngine
