+++ coreutils/src/texture.rs
@@ -4,7 +4,7 @@ use bedrock as br;
 use crate::resource::*;
 use crate::batches::TransferBatch;
 use peridot_math::Vector2;
-use std::ops::Deref;
+use std::ops::{Deref, Range};
 
 pub struct Texture2D(br::ImageView, Image);
 impl Texture2D
@@ -83,12 +83,14 @@ impl<'g> TextureInitializationGroup<'g>
 }
 impl TexturePreallocatedGroup
 {
-    pub fn alloc_and_instantiate(self, mut badget: MemoryBadget)
+    pub fn alloc_and_instantiate(self, mut badget: MemoryBadget, device: &br::Device, pd: &br::PhysicalDevice)
         -> br::Result<(TextureInstantiatedGroup, Vec<MemoryBoundResource>)>
     {
         let img_count = self.1.len();
         for isrc in self.1 { badget.add(isrc); }
-        let mut resources = badget.alloc(self.0)?;
+        let mtindex = pd.memory_properties().find_device_local_index(badget.memory_type_index_mask())
+            .expect("no matching memory type index for allocation?");
+        let mut resources = badget.alloc(device, mtindex)?;
         let textures = resources.drain(resources.len() - img_count ..)
             .map(|r| Texture2D::new(r.unwrap_image())).collect::<Result<Vec<_>, _>>()?;
 
@@ -128,78 +130,4 @@ impl Deref for TextureInstantiatedGroup
 {
     type Target = [Texture2D];
     fn deref(&self) -> &[Texture2D] { &self.1 }
-}
-
-/// Describing the type that can be used as initializer of `FixedBuffer`s
-pub trait FixedBufferInitializer
-{
-    /// Setup memory data in staging buffer
-    fn stage_data(&mut self, m: &br::MappedMemoryRange);
-    fn buffer_graphics_ready(&self, tfb: &mut TransferBatch, buf: &Buffer, range: Range<u64>);
-}
-/// The Fix-sized buffers and textures manager
-pub struct FixedMemory
-{
-    /// Device accessible buffer object
-    pub buffer: (Buffer, u64),
-    /// Host buffer staging per-frame mutable data
-    pub mut_buffer: (Buffer, u64),
-    /// The placement offset of mut_buffer data in buffer
-    pub mut_buffer_placement: u64,
-    /// Textures
-    pub textures: Vec<Texture2D>
-}
-impl FixedMemory
-{
-    /// Initialize a FixedMemory using preallocation structures
-    pub fn new<'g, I: FixedBufferInitializer + ?Sized>(
-        g: &'g Graphics,
-        mut prealloc: BufferPrealloc<'g>,
-        prealloc_mut: BufferPrealloc<'g>,
-        textures: TextureInitializationGroup<'g>,
-        initializer: &mut I, tfb: &mut TransferBatch) -> br::Result<Self>
-    {
-        let mut_buffer = prealloc_mut.build_upload()?;
-        let mut p_bufferdata_prealloc = prealloc.clone();
-        let imm_buffer_size = p_bufferdata_prealloc.total_size();
-        let mut_buffer_placement = p_bufferdata_prealloc.merge(&prealloc_mut);
-        let buffer = p_bufferdata_prealloc.build_transferred()?;
-
-        let tex_preallocs = textures.prealloc(&mut prealloc)?;
-        let stg_buffer_fullsize = prealloc.total_size();
-        let stg_buffer = prealloc.build_upload()?;
-
-        let (mut mb, mut mb_mut) = (MemoryBadget::new(g), MemoryBadget::new(g));
-        mb.add(buffer);
-        mb_mut.add(mut_buffer);
-        let (textures, mut bufs) = tex_preallocs.alloc_and_instantiate(mb)?;
-        let buffer = bufs.pop().expect("objectless").unwrap_buffer();
-        let mut_buffer = mb_mut.alloc_upload()?.pop().expect("objectless").unwrap_buffer();
-        let mut mb_stg = MemoryBadget::new(g);
-        mb_stg.add(stg_buffer);
-        let stg_buffer = mb_stg.alloc_upload()?.pop().expect("objectless").unwrap_buffer();
-
-        stg_buffer.guard_map(stg_buffer_fullsize, |m| { textures.stage_data(m); initializer.stage_data(m); })?;
-
-        textures.copy_from_stage_batches(tfb, &stg_buffer);
-        tfb.add_mirroring_buffer(&stg_buffer, &buffer, 0, imm_buffer_size);
-        initializer.buffer_graphics_ready(tfb, &buffer, 0 .. imm_buffer_size);
-
-        Ok(FixedMemory
-        {
-            buffer: (buffer, imm_buffer_size), mut_buffer: (mut_buffer, prealloc_mut.total_size()),
-            mut_buffer_placement,
-            textures: textures.into_textures()
-        })
-    }
-
-    pub fn range_in_mut_buffer<T>(&self, r: Range<T>) -> Range<T> where
-        T: std::ops::Add<Output = T> + std::convert::TryFrom<u64> + Copy
-    {
-        match T::try_from(self.mut_buffer_placement)
-        {
-            Ok(p) => r.start + p .. r.end + p,
-            Err(_) => panic!("Overflowing Placement offset")
-        }
-    }
 }
